Ensure test_multiprocess_preference_observer waits for reload to complete. (#39465)
The test could intermittently fail if the second evaluate_javascript()
command was received in the script thread before the reloaded page
actually existed. By serializing the steps so that the page is
completely loaded after the reload command is issued, we avoid this
race.

Testing: Locally reproduced, could not reproduce after the changes.
Fixes: #38920

Signed-off-by: Josh Matthews <josh@joshmatthews.net>