servoshell: Do not hide window in macOS when taking screenshot (#39500)
Screenshots for tests are taken in headless mode now, in which case the
window does not show up Dock.app. In addition, servoshell now supports
taking a screenshot in headed mode, in which case one can keep using the
browser after the screenshot happens. Hiding the window makes this
impossible. Given these two things, it seems that this functionality
isn't necessary any longer.

Testing: It's difficult to test this kind of system integration with
servoshell.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>