Add style as potential preload destination (#39549)
This was missed during the previous implementation and was the reason
that the CSP tests weren't working.

It also updates a test to ensure that audio and video are not preloaded.
No browsers do that and with this fix, the test now passes in Chrome. In
Firefox it still fails as it doesn't implement `.vtt` support.

Part of #35035

Signed-off-by: Tim van der Lippe <tvanderlippe@gmail.com>