net: Share SVG font database between all image caches in the same process (#39629)
load_system_fonts is relatively slow, since it needs to scan every font
in the system font directories. There's no need to repeat this work for
every single SVG on a page; we can do it once when creating an image
cache in a process and reuse that font information for all derived image
caches in the same process.

Testing: Manually verified that pages with SVGs still load. There should
be no observable behaviour change in WPT tests.

Signed-off-by: Josh Matthews <josh@joshmatthews.net>