webdriver: Apply script timeout to related IPC messaging as well (#39682)
In some cases scripts never call their `setTimeout` guard. In that case
also
timeout if we never hear a reply from the script thread. In general, the
script timeout is applied to all synchronous IPC waiting for script
replies.

In addition, rework timeout setting a bit to be more similar to the
specifciation.

Testing: This causes some tests run with the WebDriver test runner to
fail with
TIMEOUT rather than CRASH, accurately reflecting their reason for
failing.

---------

Signed-off-by: Martin Robinson <mrobinson@igalia.com>
Signed-off-by: Euclid Ye <yezhizhenjiakang@gmail.com>
Co-authored-by: Delan Azabani <dazabani@igalia.com>
Co-authored-by: Euclid Ye <yezhizhenjiakang@gmail.com>