fonts: Use OS/2 table to determine if font face is already bold when synthesizing bold face for FreeType platform (#39681)
The previous implementation in #39519 mistakenly used the
`FontTemplateDescriptor` to determine whether the font face itself is
already bold. This PR fixes that by using the `usWeightClass` in the
`OS/2` table of the font face to determine if the font face is bold
already.

Testing: A new testcase
`/tests/wpt/mozilla/tests/css/font_synthesis_weight_static_bold.html` is
getting added in #39633. This test checks whether a bold font face gets
"double emboldened"
Depends on: #39633

---------

Signed-off-by: Minghua Wu <michael.wu1107@gmail.com>
Signed-off-by: minghuaw <michael.wu1107@gmail.com>
Co-authored-by: Martin Robinson <mrobinson@igalia.com>