libservo: Prevent egui from consuming `PinchGesture` events on macOS (#39867)
For some reason, egui consumes all `PinchGesture` events trigged on
macOS, even if the cursor is over the WebView. This change is a
workaround which enables pinch zoom to work on Macbooks once again.

Testing: This is a fix for the input handling layer of servoshell, which
is currently untested.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>
Co-authored-by: Rakhi Sharma <atbrakhi@igalia.com>