constellation: Unload documents when traversing session history. (#39871)
We're not yet at the point where it makes sense to label HTML
specification steps in our implementation of navigation/session history
because we're so far from the specified model. However, this change
reflects a very small part of the specification, where we:
* https://html.spec.whatwg.org/#apply-the-history-step
*
https://html.spec.whatwg.org/#deactivate-a-document-for-a-cross-document-navigation
*
https://html.spec.whatwg.org/#updating-the-traversable:unload-a-document-and-its-descendants
*
https://html.spec.whatwg.org/#unloading-documents:unload-a-document-and-its-descendants
* https://html.spec.whatwg.org/#unload-a-document

We unload documents as expected when navigating, but not when traversing
the session history.

Testing: Newly passing test.
Fixes: #39798

Signed-off-by: Josh Matthews <josh@joshmatthews.net>