servoshell/desktop: condition gamepad support on pref (#39889)
In desktop servoshell, only initialize gamepad support if pref is on, in
particular because gilrs spawns background threads which we don't need
unless the pref is on(and which cannot be shutdown cleanly because they
are outside of our reach). Note that gamepad is enabled by default, so
avoiding the threads requires `dom_gamepad_enabled=false`.

Testing: Manual; gilrs threads abstent when running servo with `--pref
dom_gamepad_enabled=false`
Fixes: https://github.com/servo/servo/issues/30849

Signed-off-by: gterzian <2792687+gterzian@users.noreply.github.com>