CI: Don't run macos (x86) in full (#39968)
Only do x86 builds on macos on push and on nightly builds. Also test
building libservo on arm mac instead of x86 mac. Also fixes an issue
when merging macos arm jobs in the try parser.
The main goal of this PR is to improve CI merge times again (since the
concurrency limit of github-hosted runners, has been slowing down the MQ
a lot).

Testing: This changes a CI workflow.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>