script: Trigger a rendering update when calling `IntersectionObserver.observe` (#39967)
Observers should trigger an observation when they are installed, but
this can only happen if a rendering update happens. This change makes
sure that a rendering opportunity is triggered. This isn't discused in
the specification because we only triggering rendering updates when
necessary for performance reasons.

Testing: This fixes a timeout in
`/intersection-observer/cross-origin-tall-iframe.sub.html`.
Fixes #39831.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>