cargo-deny: add exceptions for 'unic-*' crates. (#40058)
These crates are now marked as 'unmaintained' and cause `test-tidy` to
fail on CI. They are pulled in by the `urlpattern 0.3` crate which needs
to be upgraded to `0.4` but that is blocked to the duplication of some
icu4x crates (which need to be upgraded to 2.0) and a few other crates.

Testing:  No testing needed.

Signed-off-by: Mukilan Thiyagarajan <mukilan@igalia.com>