Linux: Detect JIT support (#40071)
On some systems (specifically ohos / HarmonyOS) JIT may be forbidden,
which would cause spidermonkey to crash.
To avoid this, we test if JIT would work on startup, by remapping a page
to be executable and disabling JIT if it not allowed.
The runtime check is currently only implemented for linux, but similar
checks could easily be added for other platforms which also may have
similar restrictions.

Testing: TBD
Fixes: #40029

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>