compositing: Move WebRender initialization to `IOCompositor` creation (#40101)
In order to support multiple WebRender instances in the renderer, we
need the ability for the renderer itself to manage their creation and
destruction. This change moves the initialization of the (currently)
single WebRender instances to the renderer initialization.

Testing: This should not change behavior, so is covered by existing
tests.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>
Co-authored-by: Mukilan Thiyagarajan <mukilan@igalia.com>