script: Use interior mutability for stream controller queues. (#40138)
The stream controller code has many instances of mutably borrowing the
queue and then invoking an operation that can trigger a GC. To avoid
that, we need to make the queue mutability more granular with interior
mutability.

Testing: Many stream tests no longer crash when using a debug mozjs
build and full GC zeal. This configuration is not run in CI.
Fixes: #39947

Signed-off-by: Josh Matthews <josh@joshmatthews.net>