script: Don't bail out of "assign slottables for a tree" when the target is a slot element (#40190)
When we are unbinding a slot from the tree then "assign slottables for a
tree" is called with the slot, which unassigns all slottables that are
no longer in the slot. Without this change the slot would keep its
assigned slottables.

cc @TimvdLippe 

Testing: New web platform tests start to pass
Fixes https://github.com/servo/servo/issues/35188

---------

Signed-off-by: Simon Wülker <simon.wuelker@arcor.de>