ci: Tokenless self-hosted runner select (#39900)
one of the biggest gaps in our self-hosted runner coverage is that pull
request checks can’t run on self-hosted runners, because reserving a
runner currently requires a MONITOR_API_TOKEN secret that is unavailable
to pull requests.

this patch replaces the authenticated requests to POST
/profile/&lt;image>/take?unique_id&amp;qualified_repo&amp;run_id with
unauthenticated requests to [POST
/select-runner?unique_id&amp;qualified_repo&amp;run_id](https://github.com/servo/ci-runners/commit/05c1f3955248c0726d28c7933003834b4568bc95#diff-b0e8b5236a6e3e0caf095b712a138e85b40d207453f8a436c43de989f4d5eabeR236),
in the runner select jobs. to guard the CI monitors against abuse, the
runner select job must prove that its request was genuine by uploading a
small artifact containing the parameters of that request, which the
monitors can then
[verify](https://github.com/servo/ci-runners/commit/05c1f3955248c0726d28c7933003834b4568bc95#diff-b0e8b5236a6e3e0caf095b712a138e85b40d207453f8a436c43de989f4d5eabeR253-R286).
once the job is marked as done, the monitors will
[refuse](https://github.com/servo/ci-runners/commit/05c1f3955248c0726d28c7933003834b4568bc95#diff-b0e8b5236a6e3e0caf095b712a138e85b40d207453f8a436c43de989f4d5eabeR242-R252)
further reservation requests.

note that the old authenticated endpoint still works, so this patch can
be reverted if anything breaks, or we can turn it off with the usual
methods (force-github-hosted-runner or NO_SELF_HOSTED_RUNNERS).

Testing:
- before (38min)
<https://github.com/servo/servo/actions/runs/18547747551/job/52869138042>
- after (8min)
<https://github.com/servo/servo/actions/runs/18834479889/job/53732619984>

Fixes: part of #38141

---------

Signed-off-by: Delan Azabani <dazabani@igalia.com>