tests: Mark iframe-same-origin.html.ini as passing. (#40255)
I can't make this test fail locally, even under heavy load. It's
entirely possible that it's been fixed by the various pieces of work
we've done to integrate ResizeObserver into the event loop and improve
the accuracy of the size calculations. If it starts failing
intermittently again, we can file an issue to track it, which will give
us much better insight into how often it fails.

Signed-off-by: Josh Matthews <josh@joshmatthews.net>