wpt: Properly process the `--headless` argument (#40289)
Instead of ignoring the `--headless` argument to the WPT test runner,
properly use it to turno on headless mode in Servo when its provided.
Additionally, when more than a single test is run, turn on headless mode
automatically to preserve existing behavior.

Although this change is for the legacy test driver, this will allow the
WebDriver test runner to run properly without providing the `--headless`
argument.

Testing: This modifies the way the test harness works, but the test
harness is untested.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>