wpt: Ensure that the correct argument is read when outputting stable unexpected log (#40409)
In the review for #40369, the name of a command-line argument was
changed, but not the code that accessed the argument. This meant that
the unexpected log was never dumped, even when there are stable
unexpected results. This change fixes that.

Testing: This should ensure that the `stable-unexpected-results-linux`
artifact
is not empty when running WPT on CI.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>