ci: Extract fast checkout steps into composite action (#40234)
this patch extracts the “faster checkout” steps we use for jobs that may
run on self-hosted runners into a composite action. the action is stored
in [servo/ci-runners](https://github.com/servo/ci-runners), to avoid an
extra checkout step, and the action calls are locked to a commit, to
avoid problems when making breaking changes.

while [servo/actions](https://github.com/servo/actions) is a reasonable
place to store this action, i’m planning to document our self-hosted CI
system and potentially make it usable by other projects, and this action
is effectively part of that system. that said, actions in that repo will
undergo
[review](https://github.com/servo/ci-runners/commit/00deefd2be1ce9f8bb065157f42b37cba04780da)
as if they were in this repo.

Testing:
- self-hosted:
[windows](https://github.com/servo/servo/actions/runs/19125076579/job/54653625055#step:2:25),
[lint](https://github.com/servo/servo/actions/runs/19125076579/job/54653630737#step:2:21)
- GitHub-hosted:
[windows](https://github.com/delan/servo/actions/runs/19125003933/job/54653410264#step:2:25),
[lint](https://github.com/delan/servo/actions/runs/19125003933/job/54653410010#step:2:21)

---------

Signed-off-by: Delan Azabani <dazabani@igalia.com>