ci: Move runner-select and runner-timeout to servo/ci-runners (#40235)
continuing from #40234, this patch moves the runner-select and
runner-timeout composite actions to
[servo/ci-runners](https://github.com/servo/ci-runners), avoiding an
extra checkout step. the action calls are locked to a commit, to avoid
problems when making breaking changes.

while [servo/actions](https://github.com/servo/actions) is a reasonable
place to store this action, i’m planning to document our self-hosted CI
system and potentially make it usable by other projects, and this action
is effectively part of that system. that said, actions in that repo will
undergo
[review](https://github.com/servo/ci-runners/commit/00deefd2be1ce9f8bb065157f42b37cba04780da)
as if they were in this repo. feel free to check the actions being
adopted in this patch:
[runner-select](https://github.com/servo/ci-runners/blob/0b9edd49381b22c14e5b123f21377269864ddf54/actions/runner-select/action.yml),
[runner-timeout](https://github.com/servo/ci-runners/blob/0b9edd49381b22c14e5b123f21377269864ddf54/actions/runner-timeout/action.yml).

Testing:
- GitHub-hosted:
[full](https://github.com/delan/servo/actions/runs/19126025419),
[coverage](https://github.com/delan/servo/actions/runs/19126126210/job/54656556991)
- self-hosted:
[full](https://github.com/servo/servo/actions/runs/19126029083),
[coverage](https://github.com/servo/servo/actions/runs/19126130030/job/54656604357)

Signed-off-by: Delan Azabani <dazabani@igalia.com>