ci: Fix ohos libservoshell.so artifact path (#40468)
cargo places the shared library artifact under
`target/<target_triple>/<profile>/libservoshell.so`. The openharmony
prefix, is something we do in mach for the hap, and is not relevant for
the .so (essentially a copy-paste mistake).
Can be compared by looking at the job run of
https://github.com/servo/servo/actions/runs/19125758002/job/54655560361
which failed due to the wrong path.

Testing: Not tested
Fixes: #40467

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>